+++ documentation/tutorial/04-logic/03-else-if-blocks/app-a/App.svelte
@@ -4,8 +4,6 @@
 
 {#if x > 10}
 	<p>{x} is greater than 10</p>
-{:else if 5 > x}
-	<p>{x} is less than 5</p>
 {:else}
-	<p>{x} is between 5 and 10</p>
+	<p>{x} is between 0 and 10</p>
 {/if}