+++ packages/svelte/src/runtime/internal/Component.js
@@ -270,6 +270,13 @@ if (typeof HTMLElement === 'function') {
 						this.$$d[name] = get_custom_element_value(name, attribute.value, this.$$p_d, 'toProp');
 					}
 				}
+				// Port over props that were set programmatically before ce was initialized
+				for (const key in this.$$p_d) {
+					if (!(key in this.$$d) && this[key] !== undefined) {
+						this.$$d[key] = this[key]; // don't transform, these were set through JavaScript
+						delete this[key]; // remove the property that shadows the getter/setter
+					}
+				}
 				this.$$c = new this.$$ctor({
 					target: this.shadowRoot || this,
 					props: {