+++ packages/svelte/test/preprocess/samples/attributes-with-closing-tag/_config.js
@@ -1,0 +1,6 @@
+export default {
+	preprocess: {
+		script: ({ attributes }) =>
+			attributes.generics && attributes.generics.includes('>') ? { code: '' } : null
+	}
+};