+++ sites/svelte.dev/src/routes/(authed)/+layout.server.js
@@ -2,7 +2,7 @@ import * as session from '$lib/db/session';
 
 /** @type {import('@sveltejs/adapter-vercel').Config} */
 export const config = {
-  runtime: 'nodejs18.x' // see https://github.com/sveltejs/svelte/pull/9136
+	runtime: 'nodejs18.x' // see https://github.com/sveltejs/svelte/pull/9136
 };
 
 export async function load({ request }) {