+++ sites/svelte.dev/src/routes/auth/login/+server.js
@@ -2,7 +2,7 @@ import { redirect } from '@sveltejs/kit';
 import { client_id, oauth } from '../_config.js';
 
 export const GET = client_id
-	? ({ url }) => {
+	? /** @param {{url: URL}} opts */ ({ url }) => {
 			const Location =
 				`${oauth}/authorize?` +
 				new URLSearchParams({
@@ -21,7 +21,7 @@ export const GET = client_id
 				<p>In order to use GitHub authentication, you will need to <a target="_blank" href="https://github.com/settings/developers">register an OAuth application</a> and create a local .env file:</p>
 				<pre>GITHUB_CLIENT_ID=[YOUR_APP_ID]\nGITHUB_CLIENT_SECRET=[YOUR_APP_SECRET]\nBASEURL=http://localhost:5173</pre>
 				<p>The <code>BASEURL</code> variable should match the callback URL specified for your app.</p>
-				<p>See also <a target="_blank" href="https://github.com/sveltejs/svelte/tree/master/site#repl-github-integration">here</a></p>
+				<p>See also <a target="_blank" href="https://github.com/sveltejs/svelte/tree/svelte-4/sites/svelte.dev#repl-github-integration">here</a></p>
 			</body>
 		`,
 				{