+++ packages/svelte/CHANGELOG.md
@@ -1,5 +1,55 @@
 # svelte
 
+## 4.2.8
+
+### Patch Changes
+
+- fix: port over props that were set prior to initialization ([#9701](https://github.com/sveltejs/svelte/pull/9701))
+
+## 4.2.7
+
+### Patch Changes
+
+- fix: handle spreads within static strings ([#9554](https://github.com/sveltejs/svelte/pull/9554))
+
+## 4.2.6
+
+### Patch Changes
+
+- fix: adjust static attribute regex ([#9551](https://github.com/sveltejs/svelte/pull/9551))
+
+## 4.2.5
+
+### Patch Changes
+
+- fix: ignore expressions in top level script/style tag attributes ([#9498](https://github.com/sveltejs/svelte/pull/9498))
+
+## 4.2.4
+
+### Patch Changes
+
+- fix: handle closing tags inside attribute values ([#9486](https://github.com/sveltejs/svelte/pull/9486))
+
+## 4.2.3
+
+### Patch Changes
+
+- fix: improve a11y-click-events-have-key-events message ([#9358](https://github.com/sveltejs/svelte/pull/9358))
+
+- fix: more robust hydration of html tag ([#9184](https://github.com/sveltejs/svelte/pull/9184))
+
+## 4.2.2
+
+### Patch Changes
+
+- fix: support camelCase properties on custom elements ([#9328](https://github.com/sveltejs/svelte/pull/9328))
+
+- fix: add missing plaintext-only value to contenteditable type ([#9242](https://github.com/sveltejs/svelte/pull/9242))
+
+- chore: upgrade magic-string to 0.30.4 ([#9292](https://github.com/sveltejs/svelte/pull/9292))
+
+- fix: ignore trailing comments when comparing nodes ([#9197](https://github.com/sveltejs/svelte/pull/9197))
+
 ## 4.2.1
 
 ### Patch Changes